Redis ipaddr cache
