+++ horde/limiter.py
@@ -1,16 +1,16 @@
 from flask_limiter import Limiter
 from flask_limiter.util import get_remote_address
 from .flask import HORDE
-from . import redis as r
+from .redis_ctrl import is_redis_up, ger_limiter_url
 
 limiter = None
 # Very basic DOS prevention
-if r.is_redis_up():
+if is_redis_up():
     try:
         limiter = Limiter(
             HORDE,
             key_func=get_remote_address,
-            storage_uri=r.ger_limiter_db(),
+            storage_uri=ger_limiter_url(),
             # storage_options={"connect_timeout": 30},
             strategy="fixed-window", # or "moving-window"
             default_limits=["90 per minute"]
